requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.6
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
Deprecated==1.2.15
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
et_xmlfile==2.0.0
fastapi==0.115.5
frozenlist==1.5.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
importlib_resources==6.4.5
Jinja2==3.1.4
jiter==0.7.1
limits==3.13.0
lxml==5.3.0
MarkupSafe==3.0.2
multidict==6.1.0
numpy==2.1.3
openai==1.55.0
openai-cli==1.0.0
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==11.0.0
propcache==0.2.0
pycparser==2.22
pydantic==2.10.0
pydantic_core==2.27.0
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.17
pytz==2024.2
requests==2.32.3
retrying==1.3.4
six==1.16.0
slowapi==0.1.9
sniffio==1.3.1
starlette==0.41.3
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
wrapt==1.16.0
yarl==1.17.2
